+++ dio/lib/src/dio_event_processor.dart
@@ -13,10 +13,15 @@ class DioEventProcessor implements EventProcessor {
   static final _dioErrorType = (DioError).toString();
 
   /// This is an [EventProcessor], which improves crash reports of [DioError]s.
-  DioEventProcessor(this._options, this._maxRequestBodySize);
+  DioEventProcessor(
+    this._options,
+    this._maxRequestBodySize,
+    this._maxResponseBodySize,
+  );
 
   final SentryOptions _options;
   final MaxRequestBodySize _maxRequestBodySize;
+  final MaxResponseBodySize _maxResponseBodySize;
 
   SentryExceptionFactory get _sentryExceptionFactory =>
       // ignore: invalid_use_of_internal_member
@@ -29,9 +34,18 @@ class DioEventProcessor implements EventProcessor {
       return event;
     }
 
+    final response = _responseFrom(dioError);
+
+    Contexts contexts = event.contexts;
+    if (event.contexts.response == null) {
+      contexts = contexts.copyWith(response: response);
+    }
     // Don't override just parts of the original request.
     // Keep the original one or if there's none create one.
-    event = event.copyWith(request: event.request ?? _requestFrom(dioError));
+    event = event.copyWith(
+      request: event.request ?? _requestFrom(dioError),
+      contexts: contexts,
+    );
 
     final innerDioStackTrace = dioError.stackTrace;
     final innerDioErrorException = dioError.error as Object?;
@@ -135,6 +149,41 @@ class DioEventProcessor implements EventProcessor {
       }
     } else if (data is List<int>) {
       if (_maxRequestBodySize.shouldAddBody(data.length)) {
+        return data;
+      }
+    }
+    return null;
+  }
+
+  SentryResponse? _responseFrom(DioError dioError) {
+    final response = dioError.response;
+
+    final headers = response?.headers.map.map(
+      (key, value) => MapEntry(key, value.join('; ')),
+    );
+
+    return SentryResponse(
+      headers: _options.sendDefaultPii ? headers : null,
+      url: response?.realUri.toString(),
+      redirected: response?.isRedirect,
+      body: _getResponseData(dioError.response?.data),
+      statusCode: response?.statusCode,
+      status: response?.statusMessage,
+    );
+  }
+
+  /// Returns the request data, if possible according to the users settings.
+  /// Type checks are based on DIOs [ResponseType].
+  Object? _getResponseData(dynamic data) {
+    if (!_options.sendDefaultPii) {
+      return null;
+    }
+    if (data is String) {
+      if (_maxResponseBodySize.shouldAddBody(data.codeUnits.length)) {
+        return data;
+      }
+    } else if (data is List<int>) {
+      if (_maxResponseBodySize.shouldAddBody(data.length)) {
         return data;
       }
     }