+++ CHANGELOG.md
@@ -5,6 +5,16 @@
 ### Features
 
 * Add integration for `PlatformDispatcher.onError` ([#915](https://github.com/getsentry/sentry-dart/pull/915))
+## 6.8.0
+
+### Fixes
+
+* Missing OS context for iOS events ([#958](https://github.com/getsentry/sentry-dart/pull/958))
+* Fix: `Scope#clone` calls the Native bridges again via the `scopeObserver` ([#959](https://github.com/getsentry/sentry-dart/pull/959))
+
+### Features
+
+* Dio Integration adds response data ([#934](https://github.com/getsentry/sentry-dart/pull/934))
 
 ## 6.7.0